drivers: sdhc: allow bandwidth selection
The current implementation uses both, host and card capabilites to derive the maximum bus width to be used.
However, in cases where a MMC device is not connected to the host via shdc using the full bus width of 8 lines,
device initialization fails.
Introducing the `bus-width` property circumvents this by reducing the host bus capabilites and allowing
communication with the MMC device using 1, 4 or 8 lines.

Signed-off-by: Mourad Kharrazi <mourad.kharrazi@ithinx.io>